add extensive validation tests
this adds several thousand tests validating the structure of the document. the tests encode all of the known relationships in the structure. Comments help clarify where we are working around one or two oddball values for historic reasons.

originally investigated three different json schema libraries for the purpose but this schema is complex -- it includes several non-normal data redundancies and weird one-off dadta exceptions. The schema would have been harder to parse and maintain than this series of tests. Most importantly, this appraoch yields better error messages when a human messes up.

It's still vital to proof read any change requests. The tests capture everythign we know about now but still allow for semantic mistakes and typos that will still validate on structure and type